move js morphing from common to sick-note-extend
I don't know if this has side effects on other pages or if this is really reusable everywhere...
we have to check this sometime..
